[fenix] For https://github.com/mozilla-mobile/fenix/issues/22155 - Inactive tabs count telemetry
A quantity probe in the metrics ping means we'll loose the granularity events
provided but it will be easier to extract the values.

For reporting whether the inactive tabs feature is enabled or not we already
have the "preferences.inactive_tabs_enabled" probe so I didn't duplicate this.

UltraBlame original commit: 8000422357504d721a31e57555c65ce44e3758a0