[fenix] For https://github.com/mozilla-mobile/fenix/issues/25538 - Add a simple way to check if the default wallpaper is to be shown
A previous check for whether the persisted name of the current wallpaper is
empty made checking whether the default wallpaper is to be shown brittle.
Instead of duplicating such a check in multiple places it is now extracted in
one method that can be used in multiple places including the new telemetry.

UltraBlame original commit: bd251a16c9db03ccf5e49b40613257127adcc724