[fenix] Close https://github.com/mozilla-mobile/fenix/issues/7227: Re-enable storage maintenance call.
Re-enable storage maintenance call by introducing WorkManager worker on A-C side and consuming it from Fenix.
The work request is periodic and the repeat interval is 24h. It requires the device to be idle and not to have
low battery. This feature is available only for Nightly for now.

UltraBlame original commit: 8933ff960d4a87355bf0ce0d53634e01cdbeaa6e