Bug 1822349 - monorepo-migration: add convert-to-hg script
This script rewrites the git repo's history and converts it to
mercurial, but stops short of merging it with mozilla-central for now.

The rewrite steps are:
- replacing octopus merges with a series of 2-parent merges, since
  mercurial doesn't have octopus merges, appending " (part N)" to the
  first line of the commit message for each
- moving all files from the repository root to `mobile/android/`
- replacing references to pull request numbers in commit messages with
  their full URL

Git-cinnabar is used for the conversion to mercurial, with a custom
change to better detect file copies/moves.

UltraBlame original commit: c23ca4f3c797f9c513701f3261201f0d17ef707d