Bug 1830112 - Display full text in compose button when large font
When using the compose button and the the size of the font is
set to largest for accessibility purposes, the font should not
cut if it doesn't fit. Instead, the button accommodates the
entirety of the text.

UltraBlame original commit: a727b07a7b3d3cd884324d484386a90b93ad8c3e