Bug 1831497 - [a11y] Dictate correct hint when selecting engine
In order for talkback to correctly dictate the hint for a newly
selected search engine in the search toolbar, it required setting
it when updating the content description of the toolbar.
Previously, when updating the content description, it was using
the previous search engine hint.

UltraBlame original commit: 45db4b4a90b845b36be105d76e30cc40cbb1e5c4