Bug 1859913 - Revert shopping CFR behavior
We recently made a change to the shopping CFRs that made them harder
to dismiss. They can only be dismissed by clicking the close X or
by clicking the link that opens the bottom sheet. The issue is that
this results in the page and the UI being unresponsive. So we want
to revert back to the original behavior where they were
dismissible by clicking outside. This change currently only affects
the shopping CFR

UltraBlame original commit: 241b7377c89af44bf31b5e8168c2c563828afba5