Bug 1868358 - [a11y] "More to consider" is read as a heading
Before, "More to consider" was not read separately as a heading
title when TalkBack was used but the whole card could be
traveled as a heading and it's content was all read at once.

Now "More to consider" heading is read separately from the rest
of the content.

UltraBlame original commit: 0acba118ffc1a6cd2781fc526d65a86515b45ae4