Bug 1872971 - Move CODEOWNERS file to correct location
The CODEOWNERS file needs to be in [the root directory][0] for it to be
evaluated. Since we've merged projects, the code and team structures
have changed, so let's start off with a simple one that includes some
known basics and iteratively add more directory paths from there.

[0]: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#codeowners-file-location

UltraBlame original commit: 79370dc399e97657619608ce88030ca18896db17